c4sp M4: binary32 floats, a REPL, and c4sp as a C4KE process
Floats ride on include/c4_float.h: the reader takes the float path only
when a token has a point or exponent (the integer path must win --
binary32 cannot hold every int), arithmetic folds promote int to float
the moment a float joins, comparisons and = go through f32 when a float
is involved (so (= 1 1.0) is true and NaN != NaN, matching JS ===), and
the printer renders at 6 digits then trims, which reproduces how alisp's
JS prints simple values: 0.5, 1, 3.14. src/c4sp/lisp/floats.lisp sticks
to exactly-representable values and matches the Node oracle byte for
byte; integer division staying integer is the one documented divergence.

One library fix: f32_mul's '1 << 47' was a bare 32-bit literal, fine
under the C4 VM where every int is a word but undefined in a native
build -- c4_float.h had simply never been compiled natively before.

-i is a line-at-a-time REPL with error recovery (parse and eval errors
report and clear), usable natively, under c4m, and piped for the test.

C4KE needed nothing: c4sp.c4r already runs as the init process
(./c4m load-c4r.c -- c4ke.c4r c4sp.c4r file.lisp), which make test-c4sp
now exercises alongside the float and REPL tests.

+++ include/c4_float.h
@@ -224,8 +224,10 @@ static int f32_mul (int a, int b) {
 	er = ea + eb - F32_BIAS;
 
 	// Two 24-bit significands give a product in [2^46, 2^48). Reduce it to the
-	// 27-bit working form, whose leading bit sits at 26.
-	if (mr >= (1 << 47)) {
+	// 27-bit working form, whose leading bit sits at 26. The 1 is cast
+	// because a native build widens the int TYPE to 64 bits but a bare
+	// literal stays C's 32-bit int, making 1 << 47 undefined.
+	if (mr >= ((int)1 << 47)) {
 		mr = f32_shr_sticky(mr, 21);
 		++er;
 	} else {
